# Replace Zeus with process forking strategy