--- a/.gitignore
+++ b/.gitignore
@@ -6,8 +6,6 @@ pkg
 spec/examples.txt
 spec/internal
 tmp
-zeus.server.log
-zeus.server-start.log
 
 # Ignore Yarn stuff
 .pnp.*