--- a/gemfiles/no_rails_rspec_lt_3_10.gemfile
+++ b/gemfiles/no_rails_rspec_lt_3_10.gemfile
@@ -14,7 +14,6 @@ gem "rubocop"
 gem "syntax_tree"
 gem "syntax_tree-haml"
 gem "syntax_tree-rbs"
-gem "warnings_logger"
 gem "rspec", "~> 3.9.0"
 
 gemspec path: "../"