--- a/gemfiles/rails_6_1_rspec_lt_3_10.gemfile
+++ b/gemfiles/rails_6_1_rspec_lt_3_10.gemfile
@@ -14,7 +14,6 @@ gem "rubocop"
 gem "syntax_tree"
 gem "syntax_tree-haml"
 gem "syntax_tree-rbs"
-gem "warnings_logger"
 gem "activerecord-jdbcsqlite3-adapter", platform: :jruby
 gem "jdbc-sqlite3", platform: :jruby
 gem "net-ftp"