--- a/lib/super_diff/rspec/operation_tree_builders/collection_containing_exactly.rb
+++ b/lib/super_diff/rspec/operation_tree_builders/collection_containing_exactly.rb
@@ -36,7 +36,7 @@ module SuperDiff
         end
 
         def build_operation_tree
-          OperationTrees::Array.new([])
+          Basic::OperationTrees::Array.new([])
         end
 
         private