--- a/spec/support/models/active_record.rb
+++ b/spec/support/models/active_record.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module SuperDiff
+  module Test
+    module Models
+      module ActiveRecord
+        def self.define_tables
+          SuperDiff::Test::Models::ActiveRecord::Person.define_table
+          SuperDiff::Test::Models::ActiveRecord::ShippingAddress.define_table
+        end
+      end
+    end
+  end
+end