--- a/spec/support/models/active_record/person.rb
+++ b/spec/support/models/active_record/person.rb
@@ -6,25 +6,23 @@ module SuperDiff
       module ActiveRecord
         class Person < ::ActiveRecord::Base
           self.primary_key = 'person_id'
+
+          def self.define_table
+            ::ActiveRecord::Base
+              .connection
+              .create_table(
+                :people,
+                id: false,
+                primary_key: 'person_id',
+                force: true
+              ) do |t|
+              t.primary_key :person_id, null: false
+              t.string :name, null: false
+              t.integer :age, null: false
+            end
+          end
         end
       end
     end
-  end
-end
-
-RSpec.configure do |config|
-  config.before do
-    ActiveRecord::Base
-      .connection
-      .create_table(
-        :people,
-        id: false,
-        primary_key: 'person_id',
-        force: true
-      ) do |t|
-        t.primary_key :person_id, null: false
-        t.string :name, null: false
-        t.integer :age, null: false
-      end
   end
 end