… CHANGELOG.md
… Gemfile.lock
… bin/start-dev
… docs/contributors/architecture/structure.md
… docs/contributors/how-to-contribute.md
… gemfiles/no_rails_rspec_gte_3_10.gemfile
… spec/unit/active_record/object_inspection_spec.rb
… zeus.json