## Summary

Integration tests without Zeus are very slow because we run each example in a new process, which has to load all dependencies and frameworks – like RSpec, SuperDiff, ActiveSupport, etc – from scratch.

Zeus improved on this, but has a client-server architecture and has been proving somewhat difficult to use. If we instead load the dependencies and then run RSpec in a new [_forked_](https://man7.org/linux/man-pages/man2/fork.2.html) subprocess, we don't have to reload anything loaded before fork time, and don't have to use Zeus.

## Performance

Run on a 10-core Apple M1 Pro with 16 GB RAM.

Bench:
```
time ./bin/rspec spec/integration/rspec
```

### Before
387.17s user 146.91s system 95% cpu 9:18.77 total

### After
88.45s user 21.43s system 97% cpu 1:52.32 total

This represents a ~5x speed-up.